20201116_2

+++ drv/uart/drv_uart_st_type_A.cpp
@@ -27,12 +27,12 @@
 #include <drv/peripherals.h>
 #include <drv/uart/drv_st_uart_type_A_register.h>
 
-static unsigned long getApb2ClkFreq(void)
+static unsigned int getApb2ClkFreq(void)
 {
 	return clock.getApb2ClkFreq();
 }
 
-static unsigned long getApb1ClkFreq(void)
+static unsigned int getApb1ClkFreq(void)
 {
 	return clock.getApb1ClkFreq();
 }
@@ -222,7 +222,7 @@ extern "C"
 
 namespace drv
 {
-	Uart::Uart(USART_TypeDef *peri, void (*clockFunc)(bool en), void (*nvicFunc)(bool en), Stream *txStream, unsigned char txChannel, unsigned short priority, unsigned long (*getClockFreq)(void)) :  Drv(clockFunc, nvicFunc)
+	Uart::Uart(USART_TypeDef *peri, void (*clockFunc)(bool en), void (*nvicFunc)(bool en), Stream *txStream, unsigned char txChannel, unsigned short priority, unsigned int (*getClockFreq)(void)) :  Drv(clockFunc, nvicFunc)
 	{
 		this->set(txChannel, (void*)&(peri->TDR), priority);
 
@@ -234,9 +234,9 @@ namespace drv
 		mHead = 0;
 	}
 
-	bool Uart::init(unsigned long baud, unsigned long receiveBufferSize)
+	bool Uart::init(unsigned int baud, unsigned int receiveBufferSize)
 	{
-		unsigned long brr, clk = mGetClockFreq();
+		unsigned int brr, clk = mGetClockFreq();
 
 		if(mRcvBuf)
 			delete mRcvBuf;
@@ -286,7 +286,7 @@ namespace drv
 
 	void Uart::isr(void)
 	{
-		unsigned long sr = mPeri->ISR;
+		unsigned int sr = mPeri->ISR;
 
 		push(mPeri->RDR);
